+++ pipeline/PrimitiveDataset.py
@@ -70,7 +70,7 @@ class PrimitiveDataset(Dataset):
 
         #point cloud & point normals
         point_cloud = torch.from_numpy(point_cloud)
-        point_normals = torch.from_numpy(mesh.face_normals[face_indices]).to(dtype=torch.float32)
+        point_normals = torch.from_numpy(mesh.face_normals[face_indices])
 
         # augmentation
         point_cloud = add_gaussian_noise(point_cloud, mean=self.mean_points, std=self.std_points) #according to paper: mean = 0.0, std = 0.01
@@ -114,9 +114,9 @@ class PrimitiveDataset(Dataset):
             "decoder_input":decoder_input,
             # "decoder_mask":(decoder_input != self.PAD).unsqueeze(0).int() & causal_mask(decoder_input.size(0)).to(dtype=torch.int64), # (seq_len, 1) & (1, seq_len, seq_len)
             "target":target.to(dtype=torch.int64),
-            "point_cloud":points.to(dtype=torch.float16),
-            "quad_ratio":torch.tensor(quad_ratio, dtype=torch.float16),
-            "face_count":torch.tensor(face_count, dtype=torch.float16),
+            "point_cloud":points.to(dtype=torch.float32),
+            "quad_ratio":torch.tensor(quad_ratio, dtype=torch.float32),
+            "face_count":torch.tensor(face_count, dtype=torch.float32),
         }
 
 def causal_mask(size):
